String#matchAll should return iterable of RegExpExecArray (fixes TypeScript#36788)
