+++ types/string.prototype.matchall/package.json
@@ -1,0 +1,7 @@
+{
+    "private": true,
+    "types": "index",
+    "typesVersions": {
+        "<=5.2": { "*": ["ts5.2/*"] }
+    }
+}
