+++ types/string.prototype.matchall/ts5.2/tslint.json
@@ -1,0 +1,1 @@
+{ "extends": "@definitelytyped/dtslint/dt.json" }
